working libxmp support for SDL and Sansa ClipZip (for small files, e.g. chiptunes)

+++ lib/rbcodec/codecs/libxmp/player.h
@@ -69,8 +69,8 @@ struct retrig_control {
 #define NOTE_KEY_CUT	(1 << 8)	/* note cut with XMP_KEY_CUT event */
 #define NOTE_GLISSANDO	(1 << 9)
 
-#define IS_VALID_INSTRUMENT(x) ((uint32)(x) < mod->ins && mod->xxi[(x)].nsm > 0)
-#define IS_VALID_INSTRUMENT_OR_SFX(x) (((uint32)(x) < mod->ins && mod->xxi[(x)].nsm > 0) || (smix->ins > 0 && (uint32)(x) < mod->ins + smix->ins))
+#define IS_VALID_INSTRUMENT(x) ((uint32)(x) < (uint32)(mod->ins) && mod->xxi[(x)].nsm > 0)
+#define IS_VALID_INSTRUMENT_OR_SFX(x) (((uint32)(x) < (uint32)(mod->ins) && mod->xxi[(x)].nsm > 0) || (smix->ins > 0 && (uint32)(x) < (uint32)(mod->ins + smix->ins)))
 
 struct instrument_vibrato {
 	int phase;
